+++ cmd/collector/main.go
@@ -36,7 +36,6 @@ import (
 	ss "github.com/jaegertracing/jaeger/plugin/sampling/strategystore"
 	"github.com/jaegertracing/jaeger/plugin/storage"
 	"github.com/jaegertracing/jaeger/ports"
-	tCollector "github.com/jaegertracing/jaeger/tchannel/collector/app"
 )
 
 const serviceName = "jaeger-collector"
@@ -94,11 +93,6 @@ func main() {
 			})
 			collectorOpts := new(app.CollectorOptions).InitFromViper(v)
 			c.Start(collectorOpts)
-			tCollectorOpts := new(tCollector.Options).InitFromViper(v)
-			tc, err := tCollector.Start("jaeger-collector", tCollectorOpts, logger, c.SpanHandlers(), strategyStore)
-			if err != nil {
-				logger.Fatal("Could not start Tchannel thrift collector", zap.Error(err))
-			}
 
 			svc.RunAndThen(func() {
 				if closer, ok := spanWriter.(io.Closer); ok {
@@ -110,9 +104,6 @@ func main() {
 
 				if err := c.Close(); err != nil {
 					logger.Error("failed to cleanly close the collector", zap.Error(err))
-				}
-				if err := tc.Close(); err != nil {
-					logger.Error("failed to cleanly close tchannel collector", zap.Error(err))
 				}
 			})
 			return nil
@@ -128,7 +119,6 @@ func main() {
 		command,
 		svc.AddFlags,
 		app.AddFlags,
-		tCollector.AddFlags,
 		storageFactory.AddFlags,
 		strategyStoreFactory.AddFlags,
 	)